Task ten

+++ 0x01-variables_if_else_while/9-print_comb.c
@@ -12,7 +12,7 @@ int main(void)
 {
 	char num[] = "0, 1, 2, 3, 4, 5, 6, 7, 8, 9\n";
 
-	for (int i = 0; i < sizeof(num) - 1; i++)
+	for (size_t i = 0; i < sizeof(num) - 1; i++)
 	{
 		putchar(num[i]);
 	}
